Renamed from ScriptSystem to ScriptLoader
