+++ include/ship/Context.h
@@ -27,7 +27,7 @@ class Config;
 class ResourceManager;
 class FileDropMgr;
 class EventSystem;
-class ScriptSystem;
+class ScriptLoader;
 class Keystore;
 
 class Context {
@@ -67,7 +67,7 @@ class Context {
     std::shared_ptr<Fast::GfxDebugger> GetGfxDebugger();
     std::shared_ptr<FileDropMgr> GetFileDropMgr();
     std::shared_ptr<EventSystem> GetEventSystem();
-    std::shared_ptr<ScriptSystem> GetScriptSystem();
+    std::shared_ptr<ScriptLoader> GetScriptLoader();
     std::shared_ptr<Keystore> GetKeystore();
 
     std::string GetName();
@@ -88,7 +88,7 @@ class Context {
     bool InitWindow(std::shared_ptr<Window> window = nullptr);
     bool InitFileDropMgr();
     bool InitEventSystem();
-    bool InitScriptSystem(std::unordered_map<std::string, std::string> compileDefines = {}, int codeVersion = 1,
+    bool InitScriptLoader(std::unordered_map<std::string, std::string> compileDefines = {}, int codeVersion = 1,
                           std::string buildOptions = "-g -Wl", std::vector<std::string> includePaths = {},
                           std::vector<std::string> libraryPaths = {}, std::vector<std::string> libraries = {});
     bool InitKeystore();
@@ -111,7 +111,7 @@ class Context {
     std::shared_ptr<Fast::GfxDebugger> mGfxDebugger;
     std::shared_ptr<FileDropMgr> mFileDropMgr;
     std::shared_ptr<EventSystem> mEventSystem;
-    std::shared_ptr<ScriptSystem> mScriptSystem;
+    std::shared_ptr<ScriptLoader> mScriptLoader;
     std::shared_ptr<Keystore> mKeystore;
 
     std::string mConfigFilePath;
