+++ include/ship/scripting/ScriptLoader.h
@@ -1,24 +1,48 @@
 #pragma once
 
+#include "ship/resource/archive/Archive.h"
+#include "ship/scripting/LibraryLoader.h"
+
+#include <unordered_map>
 #include <string>
-#include <vector>
+#include <optional>
+#include <functional>
 
 namespace Ship {
-typedef void (*ScriptFunc_t)(void);
-typedef void* ScriptHandle_t;
+
+enum class SafeLevel { DISABLE_SCRIPTS, ONLY_TRUSTED_SCRIPTS, WARN_UNTRUSTED_SCRIPTS, ALLOW_ALL_SCRIPTS };
 
 class ScriptLoader {
   public:
-    ScriptLoader() : mHandle(nullptr), mTempFile("") {
+    ScriptLoader(const std::unordered_map<std::string, std::string>& compileDefines, const uint32_t codeVersion,
+                 const std::string& buildOptions, const std::vector<std::string>& includePaths,
+                 const std::vector<std::string>& libraryPaths, const std::vector<std::string>& libraries)
+        : mCodeVersion(codeVersion), mBuildOptions(buildOptions), mIncludePaths(includePaths),
+          mLibraryPaths(libraryPaths), mLibraries(libraries), mCompileDefines(compileDefines) {
     }
 
-    std::string GenerateTempFile();
-    void Init(const std::string& path);
-    void* GetFunction(const std::string& name);
-    void Unload();
+    void Compile(const std::shared_ptr<Archive>& archive);
+    void
+    CompileAll(const std::optional<std::function<void(const std::shared_ptr<Archive>&)>>& preCallback = std::nullopt,
+               const std::optional<std::function<void()>>& postCallback = std::nullopt);
+    void LoadAll();
+    void UnloadAll();
+
+    void* GetFunction(const std::string& name, const std::string& function);
+    std::vector<std::string> GetLoadersInDependencyOrder() const;
+
+    void SetSafeLevel(SafeLevel level);
 
   private:
-    ScriptHandle_t mHandle;
-    std::string mTempFile;
+    uint32_t mCodeVersion;
+    SafeLevel mSafeLevel = SafeLevel::WARN_UNTRUSTED_SCRIPTS;
+    std::string mBuildOptions = "-g -Wl";
+    std::vector<std::string> mIncludePaths;
+    std::vector<std::string> mLibraryPaths;
+    std::vector<std::string> mLibraries;
+    std::unordered_map<std::string, std::string> mCompileDefines;
+    std::unordered_map<std::string, Scripting::LibraryLoader> mLoadedScripts;
+    std::vector<std::shared_ptr<Archive>> mLoadedArchives;
 };
+
 } // namespace Ship
